Move operations tests to separate package
