+++ ksml/src/test/java/io/axual/ksml/BasicStreamRunTest.java
@@ -29,6 +29,7 @@ import io.axual.ksml.data.notation.json.JsonNotation;
 import io.axual.ksml.definition.parser.TopologyDefinitionParser;
 import io.axual.ksml.execution.ExecutionContext;
 import io.axual.ksml.generator.YAMLObjectMapper;
+import io.axual.ksml.operation.SensorData;
 import io.axual.ksml.parser.ParseNode;
 import io.axual.ksml.type.UserType;
 import io.confluent.kafka.serializers.KafkaAvroDeserializer;
