+++ ksml/src/test/java/io/axual/ksml/operation/KSMLStateStoreTest.java
@@ -1,4 +1,4 @@
-package io.axual.ksml;
+package io.axual.ksml.operation;
 
 /*-
  * ========================LICENSE_START=================================
